Release2

+++ mAdcOW.DataStructures/ViewManager.cs
@@ -5,13 +5,13 @@ using System.IO;
 using System.Threading;
 using System.Timers;
 using Winterdom.IO.FileMap;
-using Timer = System.Timers.Timer;
+using Timer=System.Timers.Timer;
 
 namespace mAdcOW.DataStructures
 {
     internal class ViewManager : IViewManager
     {
-        private const int GrowPercentage = 25;
+        private const int GrowPercentage = 20;
 
         private readonly System.Collections.Generic.Dictionary<int, DateTime> _lastUsedThread =
             new System.Collections.Generic.Dictionary<int, DateTime>();
@@ -25,7 +25,7 @@ namespace mAdcOW.DataStructures
         private long _fileSize;
         private string _fileName;
         private bool _deleteFile = true;
-
+        
         private MemoryMappedFile _map;
         private Timer _pooltimer;
 
@@ -43,7 +43,7 @@ namespace mAdcOW.DataStructures
         /// <returns></returns>
         public Stream GetView(int threadId)
         {
-            _viewLock.EnterReadLock();
+            _viewLock.EnterUpgradeableReadLock();
             try
             {
                 _lastUsedThread[threadId] = DateTime.UtcNow;
@@ -52,34 +52,25 @@ namespace mAdcOW.DataStructures
                 {
                     return s;
                 }
-            }
-            finally
-            {
-                _viewLock.ExitReadLock();
-            }
-
-            _viewLock.EnterWriteLock();
-            try
-            {
                 return AddNewViewToThreadPool(threadId);
             }
             finally
             {
-                _viewLock.ExitWriteLock();
+                _viewLock.ExitUpgradeableReadLock();
             }
         }
 
         public void Initialize(string fileName, long capacity, int dataSize)
         {
             _dataSize = dataSize;
-            _fileSize = capacity * dataSize;
+            _fileSize = capacity*dataSize;
             _fileName = fileName;
             _map = MemoryMappedFile.Create(fileName, MapProtection.PageReadWrite, _fileSize);
         }
 
         public long Length
         {
-            get { return _fileSize / _dataSize; }
+            get { return _fileSize/_dataSize; }
         }
 
         public bool EnoughBackingCapacity(long position, long writeLength)
@@ -110,13 +101,17 @@ namespace mAdcOW.DataStructures
 
         private Stream AddNewViewToThreadPool(int threadId)
         {
-            MapViewStream mvs;
-            _viewThreadPool[threadId] = mvs = _map.MapAsStream();
-
-            Trace.Write(threadId);
-            Trace.WriteLine("Adding GC memory pressure:" + mvs.Length);
-            GC.AddMemoryPressure(mvs.Length);
-            return mvs;
+            _viewLock.EnterWriteLock();
+            try
+            {
+                MapViewStream mvs;
+                _viewThreadPool[threadId] = mvs = _map.MapAsStream();
+                return mvs;
+            }
+            finally
+            {
+                _viewLock.ExitWriteLock();
+            }
         }
 
         private void EnsureBackingFile()
@@ -175,9 +170,6 @@ namespace mAdcOW.DataStructures
         {
             if (_viewThreadPool.ContainsKey(threadId))
             {
-                Trace.Write(threadId);
-                Trace.WriteLine("Removing GC memory pressure: " + _viewThreadPool[threadId].Length);
-                GC.RemoveMemoryPressure(_viewThreadPool[threadId].Length);
                 _viewThreadPool[threadId].Close();
                 _viewThreadPool.Remove(threadId);
             }
@@ -210,7 +202,7 @@ namespace mAdcOW.DataStructures
         {
             long oldSize = _fileSize;
             long newSize = oldSize + _dataSize;
-            _fileSize = (long)((float)size * _dataSize * ((100F + percentage) / 100F)); //required filesize
+            _fileSize = (long) ((float) size*_dataSize*((100F + percentage)/100F)); //required filesize
             if (_fileSize < newSize)
             {
                 _fileSize = newSize;
@@ -220,13 +212,13 @@ namespace mAdcOW.DataStructures
         public void Dispose()
         {
             Dispose(true);
+            GC.SuppressFinalize(this);
         }
 
-        protected void Dispose(bool disposing)
+        protected virtual void Dispose(bool disposing)
         {
             if (disposing || _deleteFile)
             {
-                GC.SuppressFinalize(this);
                 DisposeAllViews();
                 CloseMapFile();
             }
@@ -247,23 +239,21 @@ namespace mAdcOW.DataStructures
             {
                 if (_deleteFile)
                 {
-                    if (File.Exists(_fileName))
-                    {
-                        Trace.WriteLine("Deleting file: " + _fileName);
-                        File.Delete(_fileName);
-                    }
+                    if (File.Exists(_fileName)) File.Delete(_fileName);
                 }
             }
             catch (UnauthorizedAccessException e)
             {
                 // TODO: Handle files which for some reason didn't want to be deleted
                 Trace.WriteLine(e.Message);
+                //throw;
             }
         }
 
         private void DisposeAllViews()
         {
-            System.Collections.Generic.List<int> cleanedThreads = new System.Collections.Generic.List<int>(_viewThreadPool.Count);
+            System.Collections.Generic.List<int> cleanedThreads =
+                new System.Collections.Generic.List<int>(_viewThreadPool.Count);
             foreach (var threadPoolEntry in _viewThreadPool)
             {
                 cleanedThreads.Add(threadPoolEntry.Key);
